Fixed it?
I don't know why i needed to do this, but Jim thinks it will help.
There is a boolean that is set to false when something is reading,
eliminating the possiblility (that never existed) Someone else revert to
an older version of the code because this does nothing except make the
whole thing run slower.